.splice works as expecte for push-like operations
Added a test. A splice is not exactly like push because it acts even more
array-ish than push. That is, if you push pairs, the pairs that have keys
matching an existing will perform a hash replacement. However, since a splice is
carefully specifying which indexes are being replaced, we treat even the hash
pairs as KnottyPairs in this case.