Fix single-valued array-hash after Map refactoring
The Perl 6 code for Pairs and Hashes have been rewritten somewhat. The change
makes it so that KnottyPair is no longer required. I also have a better
understanding of the difference between a Pair and a named argument (and, I
believe, Perl 6 has improved here too since the first writing).